Adding message actionsheet
